Store master IMG and SNR extensions as float32
The Welford accumulators in BaseMasterModule._compute_stats_from_stream
run in float64 for numerical stability over hundreds of frames, but the
final stored arrays comfortably fit in float32. For a bias master the
signal is a few-ADU offset with ~5 e- read noise, so float64 buys no
science precision and only doubles disk footprint.

Downcasting IMG and SNR at the stack_frames boundary cuts master bias
L1 files from ~566 MB to ~300 MB (verified on a 20240923 evening
cluster), bringing them close to v2's 269 MB 2D files (the ~31 MB
overhead is the SNR+MASK companion extensions, which v2 didn't store).

This change applies generically to bias/dark/flat/wls masters since the
downcast happens in the shared stack_frames base method.

+++ kpfpipe/modules/masters/base.py
@@ -129,8 +129,12 @@ class BaseMasterModule:
             snr = np.zeros_like(img)
             snr[good] = np.abs(tot[good]) / np.sqrt(var[good])
 
-            l1_arrays[f'{chip}_IMG'] = img
-            l1_arrays[f'{chip}_SNR'] = snr
+            # Welford accumulators run in float64 for numerical stability;
+            # the stored master image fits comfortably in float32 (bias signal
+            # is a few-ADU offset with ~5 e- read noise) and halves the
+            # on-disk size of the IMG and SNR extensions.
+            l1_arrays[f'{chip}_IMG'] = img.astype(np.float32)
+            l1_arrays[f'{chip}_SNR'] = snr.astype(np.float32)
             l1_arrays[f'{chip}_MASK'] = good
 
         return l1_arrays
